requirements.txt
websocket_client==1.2.3
